refactor: resolve PR review feedback
