--- src/components/header/handlers.ts
@@ -1,11 +1,0 @@
-export function handleUserIconClick() {
-  console.log("User icon clicked")
-}
-
-export function handleGlobeIconClick() {
-  console.log("Globe icon clicked")
-}
-
-export function handleMoonIconClick() {
-  console.log("Moon icon clicked")
-}
