+++ src/components/header/icon-buttons.tsx
@@ -1,11 +1,11 @@
 import { FiGlobe, FiMoon, FiUser } from "react-icons/fi"
 import { NavigationMenuLink } from "@/components/ui/navigation-menu"
-import { handleGlobeIconClick, handleMoonIconClick, handleUserIconClick } from "./handlers"
+import { cn } from "@/lib/utils"
 
 const iconConfigs = [
-  { key: "user", label: "User profile", Icon: FiUser, onClick: handleUserIconClick },
-  { key: "globe", label: "Language", Icon: FiGlobe, onClick: handleGlobeIconClick },
-  { key: "moon", label: "Theme", Icon: FiMoon, onClick: handleMoonIconClick },
+  { key: "user", label: "User profile", Icon: FiUser, onClick: () => console.log("User icon clicked") },
+  { key: "globe", label: "Language", Icon: FiGlobe, onClick: () => console.log("Globe icon clicked") },
+  { key: "moon", label: "Theme", Icon: FiMoon, onClick: () => console.log("Moon icon clicked") },
 ]
 
 export const IconButtonsMobile = () => (
@@ -21,7 +21,7 @@ export const IconButtonsMobile = () => (
 export const IconButtonsDesktop = ({ removeHoverClass }: { removeHoverClass: string }) => (
   <>
     {iconConfigs.map(({ key, label, Icon, onClick }) => (
-      <NavigationMenuLink key={key} asChild className={`${removeHoverClass} p-0`}>
+      <NavigationMenuLink key={key} asChild className={cn("p-0", removeHoverClass)}>
         <button type="button" onClick={onClick} aria-label={label}>
           <Icon size={24} className="text-text-primary" />
         </button>
